pyproject.toml
[project]
name = "skimr"
version = "1.0.0"
description = "Personal reading list site backed by an LLM pipeline."
requires-python = ">=3.11"
dependencies = [
    "openai>=1.50.0",
    "youtube-transcript-api>=0.6.0",
    "yt-dlp>=2024.12.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["pipeline"]
